backend/requirements.txt
fastapi
uvicorn
pydantic
python-dateutil
openai